버서커 패시브 업뎃

+++ 01_WinMain/CaerleonSword.h
@@ -3,7 +3,7 @@
 
 class CaerleonSword :public Item
 {
-	int mValue;
+	float mValue;
 
 public:
